leave defaultConfig const, merge in new file
